Correct API patch for making moves
When communicating with the API, the API would never be notified
when a game was over. I moved the function to check if a game was
over from the UI to the API. I also added a tie game property
to store to manage which message to display when the board was
full.